refactor: move the "handler" & "signature" files into the webhooks folder
